feat: Add Terms of Use page and navigation
Introduces a new "Terms of Use" page to the website.
Updates the Header and Footer components to include navigation links for the new page.
Modifies the main App component to conditionally render the Terms of Use page based on the current view.
Updates the view state and props to accommodate the new 'terms' view type.